feat: added economy cog
